pyproject.toml
[tool.poetry]
name = "validpdfgenerator"
version = "0.1.0"
description = ""
authors = ["Sebastian Hegardt <sehe@netcompany.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
requests = "^2.28.1"
reportlab = "^4.1.0"
Pillow = "10.1.0"
numpy = "^1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
